seedling-fo_requirements.txt
fiftyone>=0.24.0
pybioclip>=0.1.0
polars>=0.20.0
pillow>=9.0.0
opencv-python>=4.8.0
torch>=2.0.0
